implement RequestToJoinForm
