+++ client/src/pages/TeamPage.js
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import {
   NavLink,
   useLoaderData,
@@ -15,6 +14,7 @@ import AcceptButton from "../components/AcceptButton";
 import DenyButton from "../components/DenyButton";
 import TeamListings from "../components/TeamListings";
 import InviteTeammateForm from "../components/InviteTeammateForm";
+import RequestToJoinForm from "../components/RequestToJoinForm";
 
 export const TeamPage = () => {
   const { team, teammates, requested, authorizedTeammates } = useLoaderData();
@@ -28,24 +28,6 @@ export const TeamPage = () => {
   const tab = searchParams.get("tab");
   const listedUsers =
     isTeammate && tab && tab.includes("requests") ? requested : teammates;
-
-  const [submissionMessage, setSubmissionMessage] = useState("");
-  const [isSuccess, setIsSuccess] = useState(false);
-
-  const handleRequest = async (e) => {
-    e.preventDefault();
-    try {
-      await axios.post(`/api/teams/${id}/teammates`, {
-        userId: authedUser.id,
-        status: "requested",
-      });
-      setIsSuccess(true);
-      setSubmissionMessage("Request sent successfully!");
-    } catch (error) {
-      setIsSuccess(false);
-      setSubmissionMessage("Request or invite already pending.");
-    }
-  };
 
   const handleAcceptRequest = async (teammate) => {
     await axios.patch(`/api/teams/${team.id}/teammates`, {
@@ -84,32 +66,7 @@ export const TeamPage = () => {
           </div>
         </div>
         <div className="flex flex-col gap-8 sm:w-1/3 h-full">
-          {isTeammate ? (
-            <InviteTeammateForm />
-          ) : (
-            <div className="flex flex-col relative rounded-sm bg-slate-100 shadow p-6">
-              <div className="flex justify-between gap-4">
-                <p className="font-semibold text-center">
-                  Join <span className="font-bold">{name}!</span>
-                </p>
-                <button
-                  className="py-1 px-2 w-1/4 bg-blue-500 hover:bg-blue-300 rounded-sm text-white text-sm"
-                  onClick={handleRequest}
-                >
-                  Request
-                </button>
-              </div>
-              {submissionMessage && (
-                <p
-                  className={`absolute bottom-1 ${
-                    isSuccess ? "text-emerald-500" : "text-red-500"
-                  } text-[10px] lg:text-xs font-bold whitespace-nowrap mb-1`}
-                >
-                  {submissionMessage}
-                </p>
-              )}
-            </div>
-          )}
+          {isTeammate ? <InviteTeammateForm /> : <RequestToJoinForm />}
           <div className="flex flex-col max-h-60 sm:max-h-max sm:w-full sm:h-2/3 rounded-sm bg-slate-100 shadow">
             <p className="relative p-3 font-bold shadow-[0_0.3px_0.3px_rgba(0,0,0,0.2)]">
               {jobField}
